Refactor schedule component
