+++ pages/chat.vue
@@ -28,42 +28,7 @@
           <p class="mt-4 text-xl text-center underline sm:text-2xl">
             Online hours:
           </p>
-          <table class="mt-4 font-serif text-xl text-center sm:text-xl">
-            <thead>
-              <tr>
-                <th class="w-1/2 px-2 py-2 border">US Central Time GMT-6</th>
-                <th class="w-1/2 px-2 py-2 border">MY Time <br />GMT+8</th>
-              </tr>
-            </thead>
-            <tbody>
-              <tr v-for="schedule in schedules" :key="schedule.id">
-                <!-- US Time -->
-                <td class="px-2 py-2 border">
-                  {{ schedule.day }} <br />
-                  {{
-                    schedule.startTime > 12
-                      ? schedule.startTime - 12
-                      : schedule.startTime
-                  }}
-                  -
-                  {{
-                    schedule.endTime > 12
-                      ? schedule.endTime - 12
-                      : schedule.endTime
-                  }}
-                  {{ schedule.endTime > 12 ? 'PM' : 'AM' }}
-                </td>
-                <!-- Malaysia Time -->
-                <td class="px-2 py-2 border">
-                  {{ convertMalaysiaSchedule(schedule).day }} <br />
-                  {{ convertMalaysiaSchedule(schedule).startTime }}
-                  -
-                  {{ convertMalaysiaSchedule(schedule).endTime }}
-                  {{ convertMalaysiaSchedule(schedule).switchAM }}
-                </td>
-              </tr>
-            </tbody>
-          </table>
+          <ScheduleTable :schedules="schedules" />
         </div>
       </div>
 
@@ -107,12 +72,14 @@
 import ChatMainIcon from '~/components/svg/chat/ChatMainIcon.vue'
 import FloatingDisclaimer from '~/components/FloatingDisclaimer.vue'
 import Email from '~/components/Email.vue'
+import ScheduleTable from '~/components/ScheduleTable.vue'
 
 export default {
   components: {
     FloatingDisclaimer,
     ChatMainIcon,
     Email,
+    ScheduleTable,
   },
   data() {
     return {
@@ -127,21 +94,21 @@ export default {
       schedules: [
         {
           id: 1,
-          day: 'Sunday',
+          day: 'Wednesday',
           startTime: 20,
           endTime: 22,
         },
         {
           id: 2,
-          day: 'Thursday',
-          startTime: 5,
-          endTime: 7,
+          day: 'Friday',
+          startTime: 3,
+          endTime: 8,
         },
         {
           id: 3,
-          day: 'Wednesday',
-          startTime: 8,
-          endTime: 10,
+          day: 'Sunday',
+          startTime: 13,
+          endTime: 14,
         },
       ],
     }
@@ -191,40 +158,6 @@ export default {
     },
     getDescriptionString() {
       return 'Our online chat is available every Saturday 7-9pm and Sunday 3-5am (US Central Time GMT-6). You can also email us at any time.'
-    },
-    convertMalaysiaSchedule(schedule) {
-      // Calculations consider Daylight saving starts
-      const dayArray = [
-        'Sunday',
-        'Monday',
-        'Tuesday',
-        'Wednesday',
-        'Thursday',
-        'Friday',
-        'Saturday',
-        'Sunday',
-      ]
-      let mytStartTime
-      if (schedule.startTime > 12) {
-        mytStartTime = schedule.startTime - 11
-      } else {
-        mytStartTime = schedule.startTime + 1
-      }
-      let mytEndTime
-      if (schedule.endTime > 12) {
-        mytEndTime = schedule.endTime - 11
-      } else {
-        mytEndTime = schedule.endTime + 1
-      }
-      return {
-        day:
-          schedule.startTime <= 10
-            ? dayArray[dayArray.indexOf(schedule.day)]
-            : dayArray[dayArray.indexOf(schedule.day) + 1],
-        startTime: mytStartTime,
-        endTime: mytEndTime,
-        switchAM: schedule.endTime >= 12 ? 'AM' : 'PM',
-      }
     },
   },
   head() {
